Update textSubtitlesRender 2/2
Solve build warning:
/__w/tsMuxer/tsMuxer/tsMuxer/textSubtitlesRender.cpp: In member function 'int text_subtitles::TextSubtitlesRender::findUnquotedStrW(const wstring&, const wstring&)':
/__w/tsMuxer/tsMuxer/tsMuxer/textSubtitlesRender.cpp:197:18: warning: overflow in implicit constant conversion [-Woverflow]
return string::npos;
^
/__w/tsMuxer/tsMuxer/tsMuxer/textSubtitlesRender.cpp:217:17: warning: overflow in implicit constant conversion [-Woverflow]
return string::npos;

+++ tsMuxer/textSubtitlesRender.h
@@ -117,7 +117,7 @@ class TextSubtitlesRender {
 		int getRepeatCnt(uint32_t* pos, uint32_t* end);
 		std::vector<std::pair<Font,std::wstring> > processTxtLine(const std::wstring& line, std::vector<Font>& fontStack);
 		int browserSizeToRealSize(int bSize, double rSize);
-		int findUnquotedStrW(const std::wstring& str, const std::wstring& substr);
+		size_t findUnquotedStrW(const std::wstring& str, const std::wstring& substr);
 	private:
 		Font m_initFont;
 	};
